backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-jose[cryptography]
PyJWT
bcrypt
pydantic[email]
crewai
litellm
chromadb
sentence-transformers
markdown
langchain-huggingface
langchain-community
langchain-text-splitters
python-dotenv